# Add new indexes and functions to improve eth_getStorageAt performance.