As mentioned in https://github.com/cerc-io/ipld-eth-server/issues/200, retrieving a value by state_leaf_key, storage_leaf_key, and block_hash is very slow.

That example query took over 153 seconds.  The biggest contributor was the lack of a proper btree index on leaf key and block number.  The chunks of the hypertable are partitioned by block_number, and without this compound index, each chunk will need to be checked (104 each in `state_cids` and `storage_cids`).  With it, we can start with the most likely chunk, and even no exact match for the key+block is found, because of the ordering of the index we will obtain the nearest candidate at similar cost.

After adding the two (leaf_key, block_number) indexes, the time drops to ~17s; significantly faster, but still too slow for general use.

```
 Planning Time: 102.504 ms
 Execution Time: 16509.402 ms
```

Part of the issue is that obtaining the block_number in a subquery like this

```
HEADER_CIDS.BLOCK_NUMBER <=
		(SELECT BLOCK_NUMBER
			FROM ETH.HEADER_CIDS
			WHERE BLOCK_HASH = '0xf11b57c121457ef927263acd7ae88729acf99f6e1a856cefc85e55ef9e4fd977')
```
means that the query planner cannot figure out the target the candidate chunks in the planning stage and has to walk through the chunks regardless.  While still ultimately aided by the new indexes, this still has the secondary effect of collecting a much larger number of potential rows for use in the JOIN, and calling `CANONICAL_HEADER_HASH` on each.

By separating out the steps inside a function, we can obtain the starting block number from the block hash first, and by sprinkling it across the query wherever it is relevant, allow the query planner to narrow down the scope significantly.

A small additional optimization is to create a version of `was_state_leaf_removed` which takes the block number as an argument, to avoid calling re-executing the subquery `SELECT block_number FROM eth.header_cids WHERE block_hash = hash`.

The combined result is shows around 250ms "cold" and 40ms "hot":

```
explain analyze select * from get_storage_at_by_hash('0xf9dc898f06b597cae6fb275f3b035d48254bd3d2cb929be172580914e9e4d94d', '0x8a35acfbc15ff81a39ae7d344fd709f28e8600b4aa8c65c6b64bfe7fe36bd19b', '0xf11b57c121457ef927263acd7ae88729acf99f6e1a856cefc85e55ef9e4fd977');
                                                          QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------
 Function Scan on get_storage_at_by_hash  (cost=0.25..10.25 rows=1000 width=77) (actual time=231.628..231.628 rows=1 loops=1)
 Planning Time: 0.083 ms
 Execution Time: 232.638 ms

###############################

explain analyze select * from get_storage_at_by_hash('0xf9dc898f06b597cae6fb275f3b035d48254bd3d2cb929be172580914e9e4d94d', '0x8a35acfbc15ff81a39ae7d344fd709f28e8600b4aa8c65c6b64bfe7fe36bd19b', '0xf11b57c121457ef927263acd7ae88729acf99f6e1a856cefc85e55ef9e4fd977');
                                                         QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------
 Function Scan on get_storage_at_by_hash  (cost=0.25..10.25 rows=1000 width=77) (actual time=41.295..41.296 rows=1 loops=1)
 Planning Time: 0.038 ms
 Execution Time: 38.982 ms
```