@dboreham 2022-11-16
> Awesome!!
> 
> Not clear to me why the existing BTREE on `leaf_key` and BRIN on `block_number` couldn’t be leveraged to perform partition pruning. If the BRIN were changed to a BTREE, could it have? Is this some weird idiosyncrasy of TimescaleDB or something fundamental to Postgres? Makes sense to me that the compound is faster since we only need to scan a single index, but doesn't make sense to me why the pruning doesn't occur at all without it.
> 
> TimescaleDB must maintain its own index on partition boundaries, so it should know how to search through them in descending order with respect to the partition key. So, even if there were no BRIN on block_number I don’t understand why it would ever search through _every_ partition when looking for a latest result (unless the record it is searching for doesn’t exist or is in the earliest partition).

I think the answer is a combination of "the compound index is required for efficient execution of the compound predicate" and "it isn't that smart".

Considering the compound predicate first: PG knows about a pattern like `table.field1 = "foo" and table.field2 <= n order by field2 desc limit 1`. It knows that when it sees that pattern (and a few related ones), it can efficiently get to the target row by performing a search on a btree index for compound key (field1:field2). This is because a btree is ordered and the search operation terminates on the key that either matches or is the highest non-matching. If PG doesn't see that compound index then it can't use its clever trick. The two separate single field indices aren't terribly useful because one only allows efficient targeting of rows that match field1, but may have any value for field2; while the other only allows finding rows with field2 < n, but potentially any value for field1. These two pieces of information don't get us to the desired target row directly, certainly with nothing like the low cost of directly searching a compound index.

Now in some scenarios we can see that PG could have done a better job than it did. For example the brin index lookup actually slows things down because most of the time there are only a few rows matching the leaf key. It could have just looked through those rows, sorted in memory to find the target row. But it didn't do that. Sometimes it's not smart enough.

In general Timescale is able to prune to the chunk set that may plausibly contain target rows. For example if you provide a predicate like `block_number < 10000`, then it will not look in chunks for block_number > 10000. However, the comment above is about a query where we provided something like: `block_number < select <some stuff>`. In that case to Timescale this at its "planning" phase, it looks like : `block_number < <unknown value>` and therefore all chunks must be checked during planning. The fact that <unknown value> is actually 10000 doesn't become apparent until execution time when it is too late to be used for chunk pruning.